requirements.txt
flask>=2.2.5
authlib>=1.3.1
python-dotenv>=1.0.1
requests>=2.31.0
python-lichess>=0.10
berserk>=0.13.2
RPi.GPIO>=0.7.1
smbus>1.0
opencv-python>1.0